add ability to send seed for io_uring meta interface
This patchs adds support for specifying seed from fio.
Although seed should be specified by the user, we just set its value to ref_tag
value in PI buffer i.e. slba, internally in fio.
Block layer code checks if the ref_tag value in PI buffer and seed are equal,
and then only the remapping is done. So in-order to avoid pain of
specifying seed correctly as command option, we just set it correctly
internally.
This is done to make the testing easier, and avoid the hassle of
calculating correct seed value based on slba for user.

Later, I will change seed to something that can be specified by user
rather than setting it internally.

Signed-off-by: Anuj Gupta <anuj20.g@samsung.com>